Version 1.0 Alpha 07 - Multiple improvements
-Some code refactored
-Appearance of List_item_mileage improved
-Confirmation dialogs added when deleting a vehicle or clearing its mileages

+++ app/src/main/java/com/macisdev/mileageapp/MileageListFragment.kt
@@ -2,21 +2,18 @@ package com.macisdev.mileageapp
 
 import android.annotation.SuppressLint
 import android.os.Bundle
-import android.text.format.DateFormat
 import android.view.*
-import android.widget.TextView
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.view.menu.MenuBuilder
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.snackbar.Snackbar
 import com.macisdev.mileageapp.databinding.FragmentMileageListBinding
 import com.macisdev.mileageapp.model.Mileage
+import com.macisdev.mileageapp.utils.Adapters
 import com.macisdev.mileageapp.viewModels.MileageListViewModel
 import java.util.*
 
@@ -44,7 +41,7 @@ class MileageListFragment : Fragment() {
 		mileageListVM.mileageListLiveData.observe(viewLifecycleOwner) {updateMileages(it)}
 
 		gui.mileagesRecyclerView.layoutManager = LinearLayoutManager(view.context)
-		val adapter = MileageAdapter()
+		val adapter = Adapters.MileageAdapter()
 		gui.mileagesRecyclerView.adapter = adapter
 
 		gui.addMileageFab.setOnClickListener {
@@ -69,16 +66,32 @@ class MileageListFragment : Fragment() {
 				findNavController().navigate(directions)
 				true
 			}
+
 			R.id.clear_mileages -> {
-				mileageListVM.clearMileages()
-				Snackbar.make(gui.root, R.string.cleared_mileages, Snackbar.LENGTH_LONG).show()
+				AlertDialog.Builder(requireContext())
+					.setTitle(R.string.clear_mileages)
+					.setMessage(R.string.action_cannot_be_undone)
+					.setPositiveButton(R.string.accept) { _, _ ->
+						mileageListVM.clearMileages()
+						Snackbar.make(gui.root, R.string.cleared_mileages, Snackbar.LENGTH_LONG).show()
+					}
+					.setNegativeButton(R.string.cancel) {_, _ -> } //Nothing to do here
+					.show()
 				true
 			}
-			R.id.delete_vehicle -> {
-				mileageListVM.deleteVehicle()
-				Snackbar.make(gui.root, R.string.deleted_vehicle, Snackbar.LENGTH_LONG).show()
 
-				findNavController().navigateUp()
+			R.id.delete_vehicle -> {
+				AlertDialog.Builder(requireContext())
+					.setTitle(R.string.delete_vehicle)
+					.setMessage(R.string.action_cannot_be_undone)
+					.setPositiveButton(R.string.accept) { _, _ ->
+						mileageListVM.deleteVehicle()
+						Snackbar.make(gui.root, R.string.deleted_vehicle, Snackbar.LENGTH_LONG).show()
+
+						findNavController().navigateUp()
+					}
+					.setNegativeButton(R.string.cancel) {_, _ -> } //Nothing to do here
+					.show()
 				true
 			}
 			else -> super.onOptionsItemSelected(item)
@@ -86,7 +99,7 @@ class MileageListFragment : Fragment() {
 	}
 
 	private fun updateMileages(mileages: List<Mileage>) {
-		val adapter = gui.mileagesRecyclerView.adapter as MileageAdapter
+		val adapter = gui.mileagesRecyclerView.adapter as Adapters.MileageAdapter
 		adapter.submitList(mileages)
 		gui.recordsCountTextView.text = mileages.size.toString()
 		gui.averageMileageTextView.text = calculateAverageMileage(mileages)
@@ -101,44 +114,4 @@ class MileageListFragment : Fragment() {
 	private fun calculateAverageMileage(mileages: List<Mileage>) = String.format(Locale.getDefault(), "%.2f",
 		mileages.sumOf { it.mileage } / mileages.size)
 
-
-	private inner class MileageAdapter : ListAdapter<Mileage, MileageAdapter.MileageViewHolder>(MileageDiffCallback) {
-
-		private inner class MileageViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-			private var currentMileage: Mileage? = null
-			val mileageContentTextView: TextView = view.findViewById(R.id.mileage_content_text_view)
-			val dateContentTextView: TextView = view.findViewById(R.id.date_content_text_view)
-			val kilometresContentTextView: TextView = view.findViewById(R.id.kilometres_content_text_view)
-			val litresContentTextView: TextView = view.findViewById(R.id.litres_content_text_view)
-			val notesTextView: TextView = view.findViewById(R.id.notes_text_view)
-
-			fun bindData(mileage: Mileage) {
-				currentMileage = mileage
-
-				mileageContentTextView.text = String.format(Locale.getDefault(), "%.2f", mileage.mileage)
-				dateContentTextView.text = DateFormat.format("dd/MM/yy", mileage.date)
-				kilometresContentTextView.text = mileage.kilometres.toString()
-				litresContentTextView.text = mileage.litres.toString()
-				notesTextView.text = mileage.notes
-			}
-		}
-
-		override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MileageAdapter.MileageViewHolder {
-			return MileageViewHolder(layoutInflater.inflate(R.layout.list_item_mileage, parent, false))
-		}
-
-		override fun onBindViewHolder(holder: MileageAdapter.MileageViewHolder, position: Int) {
-			holder.bindData(currentList[position])
-		}
-	}
-
-	object MileageDiffCallback : DiffUtil.ItemCallback<Mileage>() {
-		override fun areItemsTheSame(oldItem: Mileage, newItem: Mileage): Boolean {
-			return oldItem.id == newItem.id
-		}
-
-		override fun areContentsTheSame(oldItem: Mileage, newItem: Mileage): Boolean {
-			return oldItem == newItem
-		}
-	}
 }
